requirements.txt
openai-agents==0.0.3
toml @ file:///home/conda/feedstock_root/build_artifacts/toml_1734091811753/work
factorio-rcon-py==2.1.3
fuzzywuzzy==0.18.0

aiohttp==3.11.18
bs4==4.13.4
langchain-text-splitters==0.3.8
sentence-transformers==4.1.0
faiss==1.9.0
# crawl4ai==0.6.3